requirements.txt
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.1
azure-mgmt-core==1.6.0
azure-mgmt-storage==23.0.1
azure-storage-blob==12.26.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Flask==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
